向量数据库/milvus/demo/03zilliz_cloud_demo/requirements.txt
# Zilliz Cloud 云端向量数据库实战依赖包

# 核心依赖
pymilvus>=2.3.0
requests>=2.28.0

# 数据处理
numpy>=1.21.0
pandas>=1.5.0

# 性能监控
psutil>=5.9.0
matplotlib>=3.6.0

# 安全加密
cryptography>=3.4.8

# 配置管理
python-dotenv>=0.19.0

# 日志记录
loguru>=0.6.0

# 时间处理
python-dateutil>=2.8.0

# JSON处理
ujson>=5.0.0

# 可选：机器学习相关
scikit-learn>=1.1.0
sentence-transformers>=2.2.0

# 可选：Web框架（如果需要构建API）
fastapi>=0.85.0
uvicorn>=0.18.0

# 开发工具
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0